Add missing item combo
This was always my understanding of the "intended" way to get this item. [I asked in Discord](https://discord.com/channels/772964112908156938/822832351797182514/967872224805654588) and was told the reason this wasn't here already [was because in Rando 3 you actually couldn't reach it with the added platforms](https://discord.com/channels/772964112908156938/822832351797182514/967872788209741845). Since the platforms are now conditional based on whether you have claw or not, we can consider this check reachable with that combination of items.